refactor(#285): props.ts -> make-props.ts
