+++ packages/kotti-ui/source/kotti-banner/KtBanner.vue
@@ -23,7 +23,7 @@ import { computed, defineComponent, ref } from '@vue/composition-api'
 
 import { KtButton } from '../kotti-button'
 import { useTranslationNamespace } from '../kotti-translation/hooks'
-import { makeProps } from '../props'
+import { makeProps } from '../make-props'
 
 import { KottiBanner } from './types'
