+++ packages/kotti-ui/source/kotti-breadcrumb/KtBreadcrumb.vue
@@ -27,7 +27,7 @@
 <script lang="ts">
 import { defineComponent } from '@vue/composition-api'
 
-import { makeProps } from '../props'
+import { makeProps } from '../make-props'
 
 import { KottiBreadcrumb } from './types'
